work items: 25679, 25676, 25675, 25674
git-svn-id: https://dbentry.svn.codeplex.com/svn@62282 81e6e737-5a20-4d0f-a97c-cb01297b46e0

+++ src/Lephone.CodeGen/RailsViewGenerator.cs
@@ -143,7 +143,7 @@ namespace Lephone.CodeGen
     public IEnumerable List;
     public long ListCount;
     public int ListPageSize;
-    public int ListPageCount;
+    public long ListPageCount;
 ";
             Process(sb, vars, delegate(HtmlBuilder b)
             {
